+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Avatar: Added a `linkProps` and `linkAs` props which will be passed to the Thumbnail child.
 -   UserBlock: Added a `linkProps` and `linkAs` props which will be passed to the Thumbnail child and used to wrap the username.
+-   UserBlock: Added a `nameProps` props which will be passed to the name block.
 
 ## [2.2.0][] - 2022-01-21